Хуй знает. Много чего сделал

+++ HighFive/Client/CChat.h
@@ -41,7 +41,7 @@ public:
 	void Render();
 	void Input();
 	void AddChatMessage(std::string text, unsigned char red = 255, unsigned char green = 255, unsigned char blue = 255, unsigned char alpha = 255);
-	void AddChatMessage(std::string text, UINT64 rgba);
+	void AddChatMessage(std::string text, unsigned int rgba);
 	void AddChatMessage(std::string text, color_t color);
 	void Clear();
 	void Toggle();
